More CSS & Error Validations

+++ react-app/src/components/Photos/DeletePhoto/index.js
@@ -3,17 +3,22 @@ import { useHistory } from "react-router-dom";
 import { useModal } from "../../../context/Modal";
 import { deletePhotoThunk } from "../../../store/photoReducer";
 
-const DeletePhoto = ({individualPhoto}) => {
+const DeletePhoto = ({individualPhoto, sessionUser}) => {
     const dispatch = useDispatch();
     const history =  useHistory();
-    const { closeModal } = useModal()
+    const { closeModal } = useModal();
 
     const handleSubmit = async (e) => {
         e.preventDefault();
-        return await dispatch(deletePhotoThunk(individualPhoto.id))
-            .then(() => history.push('/photos'))
-            .then(() => closeModal())
+        const confirmed = window.confirm(`Hey ${sessionUser.first_name}! This photo will be permanently removed, are you sure?`);
+        if (confirmed) {
+            await dispatch(deletePhotoThunk(individualPhoto.id))
+                .then(() => window.alert(`${individualPhoto.title} Photo ID:${individualPhoto.id} has been deleted.`))
+                .then(() => history.push('/photos'))
+                .then(() => closeModal())
+        }
     }
+
     return (
         <div className="delete-modal-container">
             <div className="delete-pop-up">
